Refactor main.py to use collect_client_brief and remove structure_generator; add collect_client_brief implementation
